--- a/bin/benchmark-in-docker.sh
+++ b/bin/benchmark-in-docker.sh
@@ -27,13 +27,14 @@ required_tool hyperfine
 
 # Pre-build the Docker image
 if [[ -z "${SKIP_DOCKER_BUILD}" ]]; then  
-  docker build --rm -t exercism/replace-this-with-the-track-slug-test-runner .
+  docker build --rm -t exercism/swift-test-runner .
 else
   echo "Skipping docker build because SKIP_DOCKER_BUILD is set."
 fi
 
 CURRENT_PATH=${PWD}
 
-hyperfine \
-    --parameter-list slug $(find tests -maxdepth 1 -mindepth 1 -type d -printf '%f\n' | paste -sd ",") \ 
-    "SKIP_DOCKER_BUILD=true bin/run-in-docker.sh {slug} $CURRENT_PATH/tests/{slug} $CURRENT_PATH/tests/{slug}" \
+sudo hyperfine \
+    --prepare "git clean -fdqx tests/{slug}" \
+    --parameter-list slug $(find tests -maxdepth 1 -mindepth 1 -type d -printf '%f\n' | paste -sd ",") \
+    "SKIP_DOCKER_BUILD=true bin/run-in-docker.sh {slug} $CURRENT_PATH/tests/{slug} $CURRENT_PATH/tests/{slug}"