--- a/src/WarmUp/Tests/WarmUpTests/WarmUpTests.swift
+++ b/src/WarmUp/Tests/WarmUpTests/WarmUpTests.swift
@@ -1,6 +1,12 @@
-import Testing
 import Foundation
 import Numerics
+import Testing
 
-@Test("Stub Test")
-func testStub() { /* Do nothing */ }
+@testable import WarmUp
+
+let RUNALL = Bool(ProcessInfo.processInfo.environment["RUNALL", default: "false"]) ?? false
+
+@Test("rotate", .enabled(if: RUNALL))
+func testAdd() {
+  #expect(sum(2, 3) == 5, "2+3 should equal 5")
+}